Create LLM-based features
