+++ shared/embeddings/__init__.py
@@ -1,0 +1,1 @@
+"""Shared embedding utilities."""
